refactor: rearrange utility functions
