+++ scripts/mnsit_lenet_bench/eval_mnist.sh
@@ -5,13 +5,13 @@ source "$(dirname "$0")/eval_defs.sh"
 EVAL_CONFIG_NAME="mnist_lenet"
 
 benchmarks=(
-    mnist_class_detection
-    mnist_subclass_detection
-    mnist_mislabeling_detection
-    mnist_shortcut_detection
-    mnist_mixed_datasets
-    mnist_linear_datamodeling
-    #mnist_top_k_cardinality
+    #mnist_class_detection
+    #mnist_subclass_detection
+    #mnist_mislabeling_detection
+    #mnist_shortcut_detection
+    #mnist_mixed_datasets
+    #mnist_linear_datamodeling
+    mnist_top_k_cardinality
     #mnist_model_randomization
 )
